correcting_column_targeting (#70)
* Removed bug where we deleted wrong column due to index offset

* Only allow deletion of optional columns

* Fixed bug where columns where not deleted but invalid cells was still wrongly updated

* Added a tableviewer class that handles column resizing

* Fix bug in conflict resolve.

+++ src/petab_gui/controllers/table_controllers.py
@@ -170,6 +170,8 @@ class TableController(QObject):
         )
         # test: overwrite the new model as source model
         self.proxy_model.setSourceModel(self.model)
+        # change default sizing
+        self.view.table_view.reset_column_sizes()
         self.overwritten_df.emit()
 
     def append_df(self, new_df: pd.DataFrame):
@@ -210,12 +212,12 @@ class TableController(QObject):
         for row in sorted(selected_rows, reverse=True):
             if row >= self.model.rowCount() - 1:
                 continue
+            self.model.delete_row(row)
             self.logger.log_message(
                 f"Deleted row {row} from {self.model.table_type} table."
                 f" Data: {self.model.get_df().iloc[row].to_dict()}",
                 color="orange"
             )
-            self.model.delete_row(row)
         self.model.something_changed.emit(True)
 
     def add_row(self):
@@ -240,19 +242,27 @@ class TableController(QObject):
         selected_columns = get_selected(table_view, mode=COLUMN)
         if not selected_columns:
             return
-        self.model.update_invalid_cells(selected_columns, mode="columns")
+        deleted_columns = set()
         for column in sorted(selected_columns, reverse=True):
             # safely delete potential item delegates
-            column_name = self.model.get_df().columns[column]
+            allow_del, column_name = self.model.allow_column_deletion(column)
+            if not allow_del:
+                self.logger.log_message(
+                    f"Cannot delete column {column_name}, as it is a "
+                    f"required column!",
+                    color = "red"
+                )
+                continue
             if column_name in self.completers:
                 self.view.table_view.setItemDelegateForColumn(column, None)
                 del self.completers[column_name]
-            column_name = self.model.get_df().columns[column]
+            self.model.delete_column(column)
             self.logger.log_message(
                 f"Deleted column '{column_name}' from {self.model.table_type} table.",
                 color="orange"
             )
-            self.model.delete_column(column)
+            deleted_columns.add(column)
+        self.model.update_invalid_cells(deleted_columns, mode="columns")
         self.model.something_changed.emit(True)
 
     def add_column(self, column_name: str = None):
